Microsoft SQL Server に対応 (Thanks shugonta)
git-svn-id: http://svn.local/svnroot/ennichi/trunk/kaede2nd@171 1f740167-d0ba-4b02-8f16-c7307cbaec5e

+++ kaede2nd/SeasarValueType.cs
@@ -35,7 +35,7 @@ namespace kaede2nd.SeasarTypes
 
         public override void BindValue(IDbCommand cmd, string columnName, object value)
         {
-            BindValue(cmd, columnName, value, DbType.UInt32);
+            BindValue(cmd, columnName, value, DbType.Int64);
         }
 
         #endregion
@@ -59,7 +59,7 @@ namespace kaede2nd.SeasarTypes
 
         public override void BindValue(IDbCommand cmd, string columnName, object value)
         {
-            BindValue(cmd, columnName, value, DbType.UInt32);
+            BindValue(cmd, columnName, value, DbType.Int64);
         }
 
         #endregion
